change: プレゼンターの状態とロジックをinteractorへ移動させた。
