+++ iOSEngineerCodeCheck/GitHubSearch/Protocols/GitHubSearchUsecase.swift
@@ -11,13 +11,21 @@ import Foundation
 // Interactor インプット
 protocol GitHubSearchInputUsecase: AnyObject {
     var presenter: GitHubSearchOutputUsecase? { get }
-    /// API通信を行い、GitHubのデータをデータベースから取得
-    func fetch(word: String, order: Order)
+    var itemsCount: Int { get }
+    var nextOrder: Order { get }
+    func currentItem(at index: Int) -> Item
+    func viewItem(at index: Int) -> GitHubSearchViewItem
     func cancel()
+    func cancelFetchingAndResetRepository()
+    func search(word: String)
+    func fetch(word: String, order: Order)
+    func changeRepositoryItem()
 }
 
 // Interactor アウトプット
 protocol GitHubSearchOutputUsecase: AnyObject {
-    /// 取得したGitHubデータの結果をViewへ通知
-    func didFetchResult(result: Result<RepositoryItems, Error>)
+    func didFetchSuccess()
+    func didFetchError(error: Error)
+    func startLoading()
+    func viewReloadRow(at index: Int)
 }
